pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "git-ai-commit"
version = "1.0.0"
description = "AI-powered conventional commit message generator for Git"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "ProviDER", email = "updbs1@gmail.com"}
]
keywords = ["git", "commit", "conventional-commits", "ai", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
]
requires-python = ">=3.9"
dependencies = [
    "typer>=0.12.0",
    "rich>=13.7.0",
    "openai>=1.12.0",
    "ollama>=0.3.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
]

[project.scripts]
git-ai-commit = "git_ai_commit.main:run"
gac = "git_ai_commit.main:run"

[project.urls]
Homepage = "https://github.com/uyg7x/git-ai-commit"
Repository = "https://github.com/uyg7x/git-ai-commit"
Issues = "https://github.com/uyg7x/git-ai-commit/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = []

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
